Add Cost in page create\update and fix type in database

+++ database/migrations/2024_05_28_121029_create_db_product_infos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('description')->nullable();
             $table->bigInteger('status');
             $table->bigInteger('type_product')->nullable();
-            $table->bigInteger('cost')->nullable();
+            $table->decimal('cost')->nullable();
             $table->timestamps();
         });
     }
